# Implement ASCII check