--- a/setup.cfg
+++ b/setup.cfg
@@ -27,7 +27,7 @@ install_requires =
 where = src
 
 [options.package_data]
-* = py.typed, *.pyi
+* = py.typed, *.pyi, *.h
 
 [options.extras_require]
 dev =