--- a/tests/test_internal.py
+++ b/tests/test_internal.py
@@ -661,15 +661,19 @@ def test_fastq_writer_repr(tmp_path):
 class TestAsciiCheck:
     from dnaio._core import bytes_ascii_check
 
-    ASCII_STRING = "In het Nederlands komen bijzondere leestekens niet vaak voor.".encode('UTF-8')
+    ASCII_STRING = "In het Nederlands komen bijzondere leestekens niet vaak voor.".encode('ascii')
     # NON-ASCII from the German wikipedia.
-    NON_ASCII_STRING = "In späterer Zeit trat Umlaut sehr häufig analogisch ein.".encode('UTF-8')
+    NON_ASCII_STRING = "In späterer Zeit trat Umlaut sehr häufig analogisch ein.".encode('latin-1')
 
     def test_ascii(self):
         assert self.bytes_ascii_check(self.ASCII_STRING)
 
+    def test_ascii_all_chars(self):
+        assert self.bytes_ascii_check(bytes(range(128)))
+        assert not self.bytes_ascii_check(bytes(range(129)))
+
     def test_non_ascii(self):
-        assert not self.bytes_ascii_check((self.NON_ASCII_STRING))
+        assert not self.bytes_ascii_check(self.NON_ASCII_STRING)
 
     def test_ascii_lenghts(self):
         # Make sure that the function finds the non-ascii byte correctly for