fixes #40 

Couldn't resist to implement it right away.

This implements the ASCII check highlighted in #40. The performance figures:

Current main:
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.020 s ±  0.013 s    [User: 909.9 ms, System: 109.6 ms]
  Range (min … max):    1.002 s …  1.043 s    10 runs
```
Using PyUnicode_DecodeASCII
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.233 s ±  0.012 s    [User: 1.126 s, System: 0.107 s]
  Range (min … max):    1.220 s …  1.252 s    10 runs
```

~This PR~ EDIT: Outdated. The check is now run at virtually no cost, see below.
```
...
```

~~Running the plain ASCII check takes 80 ms for 5 million records. This is the same difference as I noted between PyUnicode_DecodeLatin1 and PyUnicode_DecodeASCII. Given the small cost, I think this change is worth it.~~