@rhpvorderman 2022-02-09
I checked how memchr and CPython handle using 8-bit masks. They align the memory address first. So I did that too. The check is now effectively free:
Current main:
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.002 s ±  0.021 s    [User: 883.6 ms, System: 116.7 ms]
  Range (min … max):    0.976 s …  1.037 s    10 runs
```

This PR
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.018 s ±  0.031 s    [User: 890.8 ms, System: 126.9 ms]
  Range (min … max):    0.968 s …  1.060 s    10 runs
```

That is not a significant difference, so there is virtually no cost to adding this check.

I have also been looking at the PyUnicode_DecodeASCII implementation. https://github.com/python/cpython/blob/3.9/Objects/unicodeobject.c#L7243

I have no clue why this is so much slower. It checks the length. If it is 0 it returns an empty string reference (same object in python).  If it is 1 it returns a reference to a single character pyobject. If not it creates a new string with PyUnicode_New. Then it calls the `ascii_decode` function which checks for ASCII characters and copies at the same time. It also aligns properly. This is exactly the same as what we do now, except that we do not check for empty and 1-character strings.

The only conclusion that I can come too is that this ascii_decode call is fairly expensive because:

a. it is not in-line (separate function call)
b. By combining the ascii check and copying, it does not allow specific optimisations that memcpy might have.