@rhpvorderman 2022-02-14
Hmm, I can't reproduce the 'no cost' of this. There are some compiler shennanigans going on. It seems like dnaio can be 80 ms slower just randomly if code is rearranged while semantically nothing changes. 

There seems to be a cost to this ASCII check (in the 6-24 nanoseconds per 335 bytes range).  I find it hard to determine the actual cost, and therefore also hard to determine if it is worth it.